refactor: 컨테이너 변경사항 수정
