+++ client/src/App.tsx
@@ -4,12 +4,12 @@ import { BrowserRouter, Switch, Route } from 'lib/router';
 import { MAIN_URL, ITEM_LIST_URL, ITEM_URL, AUTH_URL, SIGNIN_URL, SIGNUP_URL, ORDER_LIST_URL } from 'constants/urls';
 
 import {
-  MainPage,
   NotFoundPage,
-  ItemListPage,
-  LoginPage,
+  MainPage,
+  SigninPage,
   SignupPage,
   AuthPage,
+  ItemListPage,
   ItemDetailPage,
   MyOrderListPage,
 } from 'pages';
@@ -21,11 +21,11 @@ const App: React.FC = () => {
       <BrowserRouter>
         <Switch>
           <Route exact path={MAIN_URL} component={MainPage} />
-          <Route exact path={SIGNIN_URL} component={LoginPage} />
+          <Route exact path={SIGNIN_URL} component={SigninPage} />
           <Route exact path={SIGNUP_URL} component={SignupPage} />
-          <Route exact path={`${ITEM_URL}/:id`} component={ItemDetailPage} />
-          <Route exact path={ITEM_LIST_URL} component={ItemListPage} />
           <Route exact path={AUTH_URL} component={AuthPage} />
+          <Route exact path={ITEM_LIST_URL} component={ItemListPage} />
+          <Route exact path={`${ITEM_URL}/:id`} component={ItemDetailPage} />
           <Route exact path={ORDER_LIST_URL} component={MyOrderListPage} />
           <Route path="" component={NotFoundPage} />
         </Switch>
