+++ client/src/containers/navbar-container.tsx
@@ -6,12 +6,12 @@ import { IUser } from 'types/auth';
 import { Navbar } from 'components';
 import AuthLogoutModal from 'components/auth/logout-modal';
 
-interface HeaderProps {
+interface NavbarProps {
   displayMain?: boolean;
   isMobile: boolean;
 }
 
-const HeaderContainer: FC<HeaderProps> = ({ displayMain = false, isMobile }) => {
+const NavbarContainer: FC<NavbarProps> = ({ displayMain = false, isMobile }) => {
   const [modalVisible, setModalVisible] = useState(false);
   const { userId }: IUser = useSelector(({ auth }: RootState) => ({
     userId: auth.user.userId,
@@ -34,4 +34,4 @@ const HeaderContainer: FC<HeaderProps> = ({ displayMain = false, isMobile }) => 
   );
 };
 
-export default HeaderContainer;
+export default NavbarContainer;
