+++ client/src/containers/search-item-container.tsx
@@ -3,10 +3,10 @@ import { useQuery } from 'lib/router';
 import { useDispatch, useSelector } from 'react-redux';
 import { RootState } from 'store';
 import { getListItem } from 'store/item';
-import { ItemListWrapper } from 'components';
 import { ESortType } from 'types/item';
+import SearchItemWrapper from 'components/item/search-item/search-item-wrapper';
 
-const ItemListContainer: FC = () => {
+const SearchItemContainer: FC = () => {
   const query = useQuery();
   const [pageId, setPageId] = useState(1);
   const [sortType, setSortType] = useState(ESortType.RECOMMEND);
@@ -34,7 +34,7 @@ const ItemListContainer: FC = () => {
   }, [query, pageId, sortType, dispatch]);
 
   return (
-    <ItemListWrapper
+    <SearchItemWrapper
       items={items}
       loading={loading}
       pageCount={pageCount}
@@ -47,4 +47,4 @@ const ItemListContainer: FC = () => {
   );
 };
 
-export default ItemListContainer;
+export default SearchItemContainer;
