+++ client/src/pages/index.ts
@@ -1,8 +1,8 @@
 export { default as NotFoundPage } from './not-found-page';
 export { default as MainPage } from './main-page';
-export { default as LoginPage } from './login-page';
+export { default as SigninPage } from './signin-page';
 export { default as SignupPage } from './signup-page';
-export { default as ItemDetailPage } from './item-detail-page';
 export { default as AuthPage } from './auth-page';
 export { default as ItemListPage } from './item-list-page';
+export { default as ItemDetailPage } from './item-detail-page';
 export { default as MyOrderListPage } from './my-order-list-page';
