+++ client/src/pages/main-page.tsx
@@ -5,7 +5,8 @@ import NavbarContainer from 'containers/navbar-container';
 import SmartMenuContainer from 'containers/smart-menu-container';
 import MainItemContainer from 'containers/main-item-container';
 import SearchContainer from 'containers/search-container';
-import { Layout, Banner } from 'components';
+import { Layout } from 'components';
+import Banner from 'components/item/main-item/banner';
 
 const Wrapper = styled.article`
   width: 100%;
@@ -13,6 +14,7 @@ const Wrapper = styled.article`
   .header {
     display: flex;
     flex-direction: column;
+    margin-bottom: 60px;
 
     ${props => props.theme?.mobile} {
       flex-direction: column-reverse;
